backend/requirements.txt
Django==4.2.28
psycopg2-binary==2.9.10
Pillow==11.1.0
djangorestframework
djangorestframework-gis
django-cors-headers
google-auth
requests
celery
beautifulsoup4
redis
